Move dashboard click-behavior code into the dashboard module
Finishes the click-behaviour split (DEV-2568). PR #77658 moved the
generic halves (getDataFromClicked, clickBehaviorIsValid) into
visualizations; the remaining code is dashboard-owned and moves to the
dashboard module:

- parameters/utils/click-behavior.ts (+spec) -> dashboard/utils/
  (click-behavior targeting: getTargetsForQuestion/Dashboard,
  formatSourceForTarget, getTargetForQueryParams, canSaveClickBehavior)
- visualizations/click-actions/lib/dashboard-click-drill.js,
  lib/question-url.ts (+spec), actions/DashboardClickAction.tsx
  -> dashboard/click-behavior/ (click-behavior execution), colocating
  the existing dashboard-click-drill spec

DashboardClickAction is removed from the shared-tier mode definitions
and injected by the dashboard-rendering compositions instead:

- DefaultMode loses it; DashCardVisualization now falls back to
  getDashboardClickActionMode (dashboard/click-behavior/mode.ts), which
  mirrors getMode with the click-behavior action appended
- PublicMode loses it; PublicOrEmbeddedDashboardPage composes
  PublicDashboardMode locally (PublicOrEmbeddedQuestionView keeps plain
  PublicMode, where the action could never fire)
- EmbeddingSdkMode and EmbeddingSdkStaticMode are SDK-only, so they move
  wholesale to embedding-sdk-bundle/lib/modes/, where importing the
  dashboard module is legal; behavior is unchanged

This removes the last would-be shared->feature edges: visualizations no
longer references dashboard click-behavior code.
