+++ frontend/src/embedding-sdk-bundle/lib/modes/EmbeddingSdkStaticMode.ts
@@ -1,4 +1,4 @@
-import { DashboardClickAction } from "metabase/visualizations/click-actions/actions/DashboardClickAction";
+import { DashboardClickAction } from "metabase/dashboard/click-behavior/DashboardClickAction";
 import type { QueryClickActionsMode } from "metabase/visualizations/types";
 
 export const EmbeddingSdkStaticMode: QueryClickActionsMode = {
