+++ frontend/src/metabase/dashboard/click-behavior/dashboard-click-drill.unit.spec.js
@@ -1,5 +1,5 @@
-import { DashboardClickAction } from "../../visualizations/click-actions/actions/DashboardClickAction";
-import { getClickBehavior } from "../../visualizations/click-actions/lib/dashboard-click-drill";
+import { DashboardClickAction } from "./DashboardClickAction";
+import { getClickBehavior } from "./dashboard-click-drill";
 
 const rootClickBehavior = { type: "link", linkType: "url" };
 const metricClickBehavior = { type: "actionMenu" };
