+++ frontend/src/metabase/dashboard/components/ClickBehaviorSidebar/ClickBehaviorSidebar.tsx
@@ -9,7 +9,7 @@ import {
   useDashboardContext,
 } from "metabase/dashboard/context";
 import { isTableDisplay } from "metabase/dashboard/utils";
-import { canSaveClickBehavior } from "metabase/parameters/utils/click-behavior";
+import { canSaveClickBehavior } from "metabase/dashboard/utils/click-behavior";
 import { clickBehaviorIsValid } from "metabase/visualizations/lib/formatting/click-data";
 import { getColumnKey } from "metabase-lib/v1/queries/utils/column-key";
 import type {
