+++ frontend/src/metabase/public/containers/PublicOrEmbeddedDashboard/PublicOrEmbeddedDashboardPage/PublicOrEmbeddedDashboardPage.tsx
@@ -1,5 +1,6 @@
 import { useMemo } from "react";
 
+import { DashboardClickAction } from "metabase/dashboard/click-behavior/DashboardClickAction";
 import { PublicOrEmbeddedDashCardMenu } from "metabase/dashboard/components/DashCard/PublicOrEmbeddedDashCardMenu";
 import { DASHBOARD_DISPLAY_ACTIONS } from "metabase/dashboard/components/DashboardHeader/DashboardHeaderButtonRow/constants";
 import { useDashboardLocationSync } from "metabase/dashboard/containers/DashboardApp/use-dashboard-location-sync";
@@ -16,10 +17,16 @@ import { parseSearchQuery } from "metabase/utils/browser";
 import { isActionDashCard, isQuestionCard } from "metabase/utils/dashboard";
 import { Mode } from "metabase/visualizations/click-actions/Mode";
 import { PublicMode } from "metabase/visualizations/click-actions/modes/PublicMode";
+import type { QueryClickActionsMode } from "metabase/visualizations/types";
 import type { EntityToken } from "metabase-types/api/entity";
 
 import { usePublicEndpoints } from "../../../hooks/use-public-endpoints";
 import { PublicOrEmbeddedDashboardView } from "../PublicOrEmbeddedDashboardView";
+
+const PublicDashboardMode: QueryClickActionsMode = {
+  ...PublicMode,
+  clickActions: [DashboardClickAction],
+};
 
 const PublicOrEmbeddedDashboardPageInner = () => {
   const { location, router } = useRouter();
@@ -82,7 +89,9 @@ export const PublicOrEmbeddedDashboardPage = () => {
           parameterQueryParams={parameterQueryParams}
           cardTitled={true}
           withFooter={true}
-          getClickActionMode={({ question }) => new Mode(question, PublicMode)}
+          getClickActionMode={({ question }) =>
+            new Mode(question, PublicDashboardMode)
+          }
           navigateToNewCardFromDashboard={null}
           onError={(error) => {
             dispatch(setErrorPage(error));
