+++ frontend/src/metabase/visualizations/click-actions/modes/DefaultMode.ts
@@ -1,7 +1,6 @@
 import type { QueryClickActionsMode } from "../../types";
 import { ColumnFormattingAction } from "../actions/ColumnFormattingAction";
 import { CombineColumnsAction } from "../actions/CombineColumnsAction";
-import { DashboardClickAction } from "../actions/DashboardClickAction";
 import { ExtractColumnAction } from "../actions/ExtractColumnAction";
 import { HideColumnAction } from "../actions/HideColumnAction";
 import { NativeQueryClickFallback } from "../actions/NativeQueryClickFallback";
@@ -12,7 +11,6 @@ export const DefaultMode: QueryClickActionsMode = {
   clickActions: [
     HideColumnAction,
     ColumnFormattingAction,
-    DashboardClickAction,
     ExtractColumnAction,
     CombineColumnsAction,
   ],
