+++ frontend/src/metabase/visualizations/click-actions/modes/PublicMode.ts
@@ -1,8 +1,7 @@
 import type { QueryClickActionsMode } from "../../types";
-import { DashboardClickAction } from "../actions/DashboardClickAction";
 
 export const PublicMode: QueryClickActionsMode = {
   name: "public",
   hasDrills: false,
-  clickActions: [DashboardClickAction],
+  clickActions: [],
 };
